orders/requirements.txt
Django==3.2.10
pytz==2021.3
psycopg2-binary==2.9.3
djangorestframework==3.13.1
python-decouple~=3.6

PyYAML~=6.0
pip~=22.0.4
setuptools~=60.9.3

requests~=2.27.1

ujson~=5.1.0

model_bakery~=1.5.0
coverage~=6.4
